src/requirements.txt
PySide6==6.8.1
google-genai==0.3.0
python-dotenv==1.0.1